% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.RiboMethSeq.R
\docType{data}
\name{RNAmodR.RiboMethSeq-datasets}
\alias{RNAmodR.RiboMethSeq-datasets}
\alias{msrms}
\title{Example data in the RNAmodR.RiboMethSeq package}
\format{a \code{ModSetRiboMethSeq} instance}
\usage{
data(msrms)
}
\description{
This contains an example ModifierSet object of type ModSetRiboMethSeq
}
\keyword{datasets}
