% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.RiboMethSeq.R
\docType{package}
\name{RNAmodR.RiboMethSeq-package}
\alias{RNAmodR.RiboMethSeq-package}
\title{RNAmodR.RiboMethSeq: Detection of 2'-O methylations by RiboMethSeq}
\description{
RNAmodR.RiboMethSeq implements the detection of 2'-O methylations on RNA from experimental data generated with the RiboMethSeq protocol. The package builds on the core functionality of the RNAmodR package to detect specific patterns of the modifications in high throughput sequencing data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FelixErnst/RNAmodR.RiboMethSeq}
  \item Report bugs at \url{https://github.com/FelixErnst/RNAmodR.RiboMethSeq/issues}
}

}
\author{
\strong{Maintainer}: Felix G.M. Ernst \email{felix.gm.ernst@outlook.com} (\href{https://orcid.org/0000-0001-5064-0928}{ORCID})

Other contributors:
\itemize{
  \item Denis L.J. Lafontaine \email{denis.lafontaine@ulb.ac.be} [contributor, funder]
}

}
\keyword{internal}
