% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.RiboMethSeq.R
\name{RNAmodR.RiboMethSeq}
\alias{RNAmodR.RiboMethSeq}
\title{RNAmodR.RiboMethSeq}
\description{
`RNAmodR.RiboMethSeq` implements the detection of 2'-O methylations from 
RiboMethSeq data using the workflow and class the package `RNAmodR` provides.
}
\seealso{
Further details are described in the man pages of the 
\code{\link[RNAmodR:Modifier-class]{Modifier}} object and the vignettes.
}
\author{
Felix G M Ernst [aut], Denis L J Lafontaine [fnd]
}
