% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modifier-RiboMethSeq-class.R,
%   R/Modifier-RiboMethSeq-viz.R
\docType{methods}
\name{ModRiboMethSeq-functions}
\alias{ModRiboMethSeq-functions}
\alias{aggregate}
\alias{modify}
\alias{settings}
\alias{visualizeData}
\alias{visualizeDataByCoord}
\alias{settings<-,ModRiboMethSeq-method}
\alias{aggregate,ModRiboMethSeq-method}
\alias{modify,ModRiboMethSeq-method}
\alias{visualizeDataByCoord,ModRiboMethSeq,GRanges-method}
\alias{visualizeData,ModRiboMethSeq-method}
\alias{visualizeDataByCoord,ModSetRiboMethSeq,GRanges-method}
\alias{visualizeData,ModSetRiboMethSeq-method}
\title{Functions for ModRiboMethSeq}
\usage{
\S4method{settings}{ModRiboMethSeq}(x) <- value

\S4method{aggregate}{ModRiboMethSeq}(x, force = FALSE)

\S4method{modify}{ModRiboMethSeq}(x, force = FALSE)

\S4method{visualizeDataByCoord}{ModRiboMethSeq,GRanges}(x, coord,
  type = c("ends", "scoreA", "scoreB", "scoreRMS", "scoreMean"),
  window.size = 15L, ...)

\S4method{visualizeData}{ModRiboMethSeq}(x, name, from = 1L, to = 30L,
  type = c("ends", "scoreA", "scoreB", "scoreRMS", "scoreMean"), ...)

\S4method{visualizeDataByCoord}{ModSetRiboMethSeq,GRanges}(x, coord,
  type = c("scoreRMS", "ends", "scoreA", "scoreB", "scoreMean"),
  window.size = 15L, ...)

\S4method{visualizeData}{ModSetRiboMethSeq}(x, name, from = 1L,
  to = 30L, type = c("scoreRMS", "ends", "scoreA", "scoreB",
  "scoreMean"), ...)
}
\arguments{
\item{x}{a \code{\link[RNAmodR:Modifier-class]{Modifier}} or a
\code{\link[RNAmodR:ModifierSet-class]{ModifierSet}} object. For more
details see also the man pages for the functions mentioned below.}

\item{value}{See \code{\link[RNAmodR:Modifier-class]{settings}}}

\item{force}{See \code{\link[RNAmodR:aggregate]{aggregate}}}

\item{coord, name, from, to, type, window.size, ...}{See 
\code{\link[RNAmodR:visualizeData]{visualizeData}}}
}
\description{
All of the functions of \code{\link[RNAmodR:Modifier-class]{Modifier}} and
the \code{\link[RNAmodR:ModifierSet-class]{ModifierSet}} classes are
inherited by the \code{ModRiboMethSeq} and \code{ModSetRiboMethSeq} classes.
}
\details{
\code{ModRiboMethSeq} specific arguments for 
\code{\link[RNAmodR:visualizeData]{visualizeData}}:
\itemize{
\item{\code{colour} - }{a named character vector of \code{length = 4} 
for the colours of the individual histograms. The names are expected to be 
\code{c("ends","scoreA","scoreB","scoreRMS","scoreMean")}}
}
}
