% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modifier-RiboMethSeq.R
\docType{class}
\name{ModRiboMethSeq}
\alias{ModRiboMethSeq}
\alias{RiboMethSeq}
\alias{ModRiboMethSeq-class}
\alias{ModRiboMethSeq,character-method}
\alias{ModRiboMethSeq,BamFileList-method}
\alias{ModRiboMethSeq,SequenceData-method}
\alias{modify,ModRiboMethSeq-method}
\title{ModRiboMethSeq}
\usage{
\S4method{ModRiboMethSeq}{character}(x, fasta, gff, modifications = NULL,
  ...)

\S4method{ModRiboMethSeq}{BamFileList}(x, fasta, gff,
  modifications = NULL, ...)

\S4method{ModRiboMethSeq}{SequenceData}(x, modifications = NULL, ...)

\S4method{modify}{ModRiboMethSeq}(x, ...)
}
\description{
title
}
