% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modifier-RiboMethSeq-class.R,
%   R/Modifier-RiboMethSeq-viz.R
\docType{class}
\name{ModRiboMethSeq}
\alias{ModRiboMethSeq}
\alias{RiboMethSeq}
\alias{ModSetRiboMethSeq}
\alias{ModRiboMethSeq-class}
\alias{modify,ModRiboMethSeq-method}
\alias{ModSetRiboMethSeq-class}
\alias{visualizeData}
\alias{visualizeDataByCoord,ModRiboMethSeq,GRanges-method}
\alias{visualizeData,ModRiboMethSeq-method}
\alias{visualizeDataByCoord,ModSetRiboMethSeq,GRanges-method}
\alias{visualizeData,ModSetRiboMethSeq-method}
\title{ModRiboMethSeq}
\usage{
\S4method{settings}{ModRiboMethSeq}(x) <- value

ModRiboMethSeq(x, annotation = NA, sequences = NA, seqinfo = NA, ...)

\S4method{aggregate}{ModRiboMethSeq}(x, force = FALSE)

\S4method{modify}{ModRiboMethSeq}(x, force = FALSE)

ModSetRiboMethSeq(x, annotation = NA, sequences = NA, seqinfo = NA)

\S4method{visualizeDataByCoord}{ModRiboMethSeq,GRanges}(x, coord,
  type = c("ends", "scoreA", "scoreB", "scoreRMS"), window.size = 15L,
  ...)

\S4method{visualizeData}{ModRiboMethSeq}(x, name, from = 1L, to = 30L,
  type = c("ends", "scoreA", "scoreB", "scoreRMS"), ...)

\S4method{visualizeDataByCoord}{ModSetRiboMethSeq,GRanges}(x, coord,
  type = c("scoreRMS", "ends", "scoreA", "scoreB"), window.size = 15L,
  ...)

\S4method{visualizeData}{ModSetRiboMethSeq}(x, name, from = 1L,
  to = 30L, type = c("scoreRMS", "ends", "scoreA", "scoreB"), ...)
}
\arguments{
\item{...}{Optional arguments overwriting default values, which are
\itemize{
\item{weights:}{The weights used for calculating the scores B and RMS 
(default: \code{weights = c(0.9,1,0,1,0.9)})}
\item{flankingRegion:}{The size of the flanking region used for calculation 
of score A as an integer value (default: \code{flankingRegion = 6L})}
\item{minSignal:}{The minimal singal at the position as integer value 
(default: \code{minSignal = 10L}). If the reaction is very specific a lower
value and even 0L may need to be used}
\item{minScoreA:}{minimum for score A to identify 2'-O methylated positions 
de novo (default: \code{minScoreA = 0.6})}
\item{minScoreB:}{minimum for score B to identify 2'-O methylated positions 
de novo (default: \code{minScoreB = 3.0})}
\item{minScoreRMS:}{minimum for score RMS to identify 2'-O methylated 
positions de novo (default: \code{minScoreRMS = 0.75})}
\item{scoreOperator:}{how the minimal score should be used as logical 
operator. "&" requires all minimal values to be exceeded, whereas "|" detects
positions, if at least one minimal values is exceeded (default: 
\code{scoreOperator = "|"})}
\item{maxLength:}{The default read length. Reads with this length or longer
are discarded, since they represent non-fragemented reads. This might need to
be adjusted for individual samples dending on the experimental conditions.
This is argument is passed on to 
\code{\link[RNAmodR:ProtectedEndSequenceData]{ProtectedEndSequenceData}}
(default: \code{maxLength = 50L})}
\item{other arguments}{which are passed on to 
\code{\link[RNAmodR:ProtectedEndSequenceData]{ProtectedEndSequenceData}}}
}}
}
\description{
title

The score MAX as described by publications from the Motorin lab are not 
implemented since an unambigeous description is not available from the 
literature.

To disable minimal values for modification calling, set them to \code{Inf}
}
\details{
\code{ModRiboMethSeq} specific arguments for \link{visualizeData}:
\itemize{
\item{\code{colour} - }{a named character vector of \code{length = 4} 
for the colours of the individual histograms. The names are expected to be 
\code{c("ends","scoreA","scoreB","scoreRMS")}}
}
}
\author{
Felix G.M. Ernst \email{felix.gm.ernst@outlook.com}
}
